django-first/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.2
distlib==0.3.9
Django==5.2.1
django-cors-headers==4.7.0
django-oauth-toolkit==3.0.1
djangorestframework==3.16.0
filelock==3.18.0
idna==3.10
jwcrypto==1.5.6
oauthlib==3.2.2
packaging==25.0
pipenv==2025.0.2
platformdirs==4.3.8
psycopg2-binary==2.9.10
pycparser==2.22
requests==2.32.3
setuptools==80.4.0
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
virtualenv==20.31.2
